@@ -5,8 +5,8 @@
   "dependencies": {
     "@emotion/core": "^10.0.6",
     "@emotion/styled": "^10.0.6",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
+    "react": "^16.8.0-alpha.1",
+    "react-dom": "^16.8.0-alpha.1",
     "react-outside-click-handler": "^1.2.2",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
