@@ -1,13 +1,15 @@
 import React from 'react';
-import { BrowserRouter as Router } from 'react-router-dom';
-import TestRenderer from 'react-test-renderer';
+import { MemoryRouter as Router } from 'react-router-dom';
+import ReactDOM from 'react-dom';
 import Nav from '../Nav';
 
 it('renders correctly', () => {
-  const tree = TestRenderer.create(
+  const div = document.createElement('div');
+  ReactDOM.render(
     <Router>
       <Nav />
-    </Router>
-  ).toJSON();
-  expect(tree).toMatchSnapshot();
+    </Router>,
+    div
+  );
+  ReactDOM.unmountComponentAtNode(div);
 });
